backend/requirements.txt
bs4
snscrape
playwright
requests
praw
pytrends
qdrant_client
sentence_transformers
google-generativeai
python-dotenv
chromadb
tweepy